requirements.txt
azure-identity
azure-mgmt-subscription
azure-mgmt-compute
azure-mgmt-monitor
python-dotenv